tg-6807 [BE] Agrega lógica para obtener los distintos tipos de vías y

See merge request yaerobi/HistorialClinica-LaRioja!18

+++ back-end/clinic-history/src/main/java/ar/lamansys/sgh/clinichistory/infrastructure/output/repository/ips/QuantityRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface QuantityRepository extends JpaRepository<Quantity, Integer> {
+public interface QuantityRepository extends JpaRepository<Quantity, Long> {
 
 }
